Running dotenv before any lanes called so disable color can be applied to everything
