+++ fastlane/lib/fastlane/lane_manager.rb
@@ -49,7 +49,7 @@ module Fastlane
       # Setting this environment variable causes xcodeproj to work around the problem
       ENV["FORK_XCODE_WRITING"] = "true" unless platform == 'android'
 
-      load_dot_env(env)
+      Fastlane::Helper::DotenvHelper.load_dot_env(env)
 
       started = Time.now
       e = nil
