+++ fastlane_core/lib/fastlane_core/ui/implementations/shell.rb
@@ -26,8 +26,6 @@ module FastlaneCore
         "#{format_string(datetime, severity)}#{msg}\n"
       end
 
-      require 'fastlane_core/ui/disable_colors' if FastlaneCore::Helper.colors_disabled?
-
       @log
     end
